requirements.txt
langchain_google_genai
load_dotenv
crewai
crewai_tools